Done Cache

+++ app/Services/Brands/BrandService.php
@@ -26,7 +26,7 @@ class BrandService
         }
         return $brand;
     }
-    public function getBrandsForDataTales()
+    public function getBrandsForDataTables()
     {
         $brands = $this->brandRepository->getBrands();
 
@@ -63,7 +63,7 @@ class BrandService
     {
         $brand = self::getBrandById($id);
 
-        if ($data['logo'] != null) {
+        if (array_key_exists('logo', $data) && $data['logo'] != null) {
             $this->imageManager->deleteImageFromLocal($brand->logo);
             $fileName = $this->imageManager->uploadSingleImage($data['logo'], '/', 'brands');
             $data['logo'] = $fileName;
